refactor(togglecontent): move to components folder
