+++ site/src/components/ToggleContent.jsx
@@ -1,0 +1,25 @@
+import React, { useRef } from "react";
+
+const ToggleContent = ({ children, showState }) => {
+  const childRef = useRef();
+
+  return (
+    <div
+      className="detail"
+      ref={childRef}
+      style={
+        showState
+          ? {
+              height: childRef.current && childRef.current.scrollHeight + "px",
+            }
+          : {
+              height: "0px",
+            }
+      }
+    >
+      {children}
+    </div>
+  );
+};
+
+export default ToggleContent;
