--- site/src/components/hooks/toggleContent.js
@@ -1,24 +1,0 @@
-import React, {useRef} from 'react';
-
-const ToggleContent = ({ children , showState}) => {
-    const childRef = useRef();
-    
-    return (
-        <div 
-            className="detail"
-            ref={childRef}
-            style={
-                    showState ? {
-                            height: childRef.current && childRef.current.scrollHeight + "px",
-                            }
-                        : {
-                            height: "0px",
-                            }
-                }
-        >
-          {children}
-        </div>
-    );
-};
-
-export default ToggleContent;
